Centralise exceptions.
